debugd: fix local bug in systrace.sh script
Replace the use of 'local' keyword in the 'start)' case block with an
equivalent setting and unsetting of the variable.

The local keyword can only be used within a function context, more
compliant shells reject the use within a case block with the error:

  "local: not in a function"

Versions of dash prior to version 0.5.9.1-r3 silently ignored the use of
the local keyword, but with the upgrade to this version introduced on
Nov 10 2017 in the chromiumos/overlays/portage-stable project (d0ec70a4d
- "dash: version bump"), the script stopped working.  The collection of
system tracing information on ChromeOS devices through devtools has been
broken since then.

BUG=chromium:803813
TEST=emerge chromeos-base/debugd and deploy to DUT
TEST=on DUT, run /usr/libexec/debugd/helpers/systrace.sh start
TEST=confirm "local: not in a function" message doesn't appear
TEST=run /usr/libexec/debugd/helpers/systrace.sh stop
TEST=confirm ftrace data is dumped to stdout

Change-Id: Ifdb4dbf8f0696489df53a408ac1e8ee812fa4fb6
Reviewed-on: https://chromium-review.googlesource.com/888923
Commit-Ready: Stephen Kyle <stephen.kyle@arm.com>
Tested-by: Stephen Kyle <stephen.kyle@arm.com>
Reviewed-by: oysteine <oysteine@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>