cryptohome: fix fragility in MountCryptohomeInner
This code only does what is intended because the
MOUNT_ERROR_FATAL value is defined (elsewhere, in system_api)
as 1, and so it lines up with the standard integral values for
booleans. Use logical and instead of bitwise and so that even
if the value changes this doesn't break.

While we're here, rename the recreate_decrypt_fatal argument to
make it slightly more clear.

BUG=None
TEST=unit tests

Change-Id: I3f30cdf72a89d35b121cdc27cd752231abc55c42
Reviewed-on: https://chromium-review.googlesource.com/898000
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>