cryptohome: remove Mount D-Bus call and AsyncMount method
AsyncMount users have all moved to MountEx. Service::Mount is
still around for the time being because it's used by stateful
recovery and Service::MountEx is tied to D-Bus bindings enough
that moving it over would be a much bigger deal than is in
scope for this patch.

However, since there are no remaining external users of Mount or
AsyncMount, we can get rid of the D-Bus endpoints and interface
functions, so it does still allow us to simplify somewhat.

CQ-DEPEND=CL:1000693
BUG=chromium:765525
TEST=unit tests

Change-Id: Ie38e01dbbcf2d908896a8f9911f2d949b3e531a5
Reviewed-on: https://chromium-review.googlesource.com/900283
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>