cryptohome: remove AddKey and AsyncAddKey
The only user of AsyncAddKey is SupervisedUserAuthenticator in
Chrome, which is dead code and is being removed in CL:898252.
There are also no users of either AddKey or AsyncAddKey via CLI
or D-Bus in Chrome OS.

This also allows us to get rid of MountTaskAddPasskey.

BUG=chromium:809616
TEST=unit tests

Change-Id: Ibe6f9d1e7861c271daa6cbff81603bc333c6007c
Reviewed-on: https://chromium-review.googlesource.com/899507
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>