authpolicy: Rename an error enum
We decided to get rid of Windows policy in the authpolicy daemon and
instead replicate the one Windows policy we use so far as device policy.
Thus, user policy fetch now relies on the availability of device policy,
not Windows policy, but it's still the same error.

Keeps the old error until Chrome uses the new one.

BUG=chromium:807999
TEST=cros_run_unit_tests --board=amd64-generic --packages authpolicy

Change-Id: I6e9b30b13532a8e601e66847e7d944b1d3859d41
Reviewed-on: https://chromium-review.googlesource.com/901285
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>