chromeos-config: Implement yaml as a subclass of CrosConfigImpl
We have everything ready now to switch the yaml implementation over to
use the same base class as the fdt implementation.

Rewrite CrosConfigJson so that is a simple subclass of the base
implementation. Enable running the tests for yaml (as well as the existing
fdt).

This completes the basic converion of the host tools to support yaml.

Some work still remains (not a complete list):
- Adjust the yaml schema to adher to the v2 ideal (some of which is TBD)
- Migrate existing boards over (using dts2yaml, etc.)
- Implement SKU matching, after discussion about how best to do this in
    yaml
- Consider (again) allowing the validator to support validation of dict
    names
- Revert to the stricter yaml validation (e.g. required properties)
- Clean out some fdt features which are not really needed in yaml
- Try to make the yaml validator more user-friendly (e.g. sometimes it
    dumps the entire config when it sees an error)

BUG=chromium:796983
TEST=./run_tests.sh

Change-Id: Ifaa13ef316992c7e9234fe5b5c81007bc83a6745
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/902937
Reviewed-by: Jason Clinton <jclinton@chromium.org>