cryptohome: Add protobuf for hash tree leaf data
Instead of writing a raw vector, encapsulate the leaf data
into a protobuf before writing it out to the disk.

Also, modify the logic of GenerateAndStoreHashCache(),
so that all the leaf nodes are first updated in the
HashCache file, and then CalculateHash() generates all
the inner hashes.

Additonally, slightly change the semantics of
GetLabelData(); now, if the label doesn't exist
in the tree, return an all-zero HMAC. Return false
only if we aren't able to read the leaf label value
due to an error.

Also, change the return types of PersistentLookupTable
functions to disambiguate between table storage errors,
and keys not being present.

Finally, remove some unused code in the unit tests.

BUG=chromium:794010, chromium:809703
TEST=cryptohome unit tests pass

Change-Id: I5dd4fe37df8dbd1b34bdc40e2e42a64e96b4233c
Reviewed-on: https://chromium-review.googlesource.com/902612
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Prashant Malani <pmalani@google.com>