apk-cache: use ScopedTempDir::GetPath
Upstream versions of libchrome have gotten rid of ScopedTempDir's
inline path method because they wanted to introduce a DCHECK that
helps make sure people create the temp directory before using it.
To avoid introducing a header dependency on the logging header
they took this method out of the header so it wouldn't be inlined.

BUG=b:37434548
TEST=unit tests

Change-Id: I6352a6600bffacf673195e9e446d1ebbafbc4b8d
Reviewed-on: https://chromium-review.googlesource.com/905452
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>