arc-setup: use ScopedTempDir::GetPath
Upstream versions of libchrome have gotten rid of ScopedTempDir's
inline path method because they wanted to introduce a DCHECK that
helps make sure people create the temp directory before using it.
To avoid introducing a header dependency on the logging header
they took this method out of the header so it wouldn't be inlined.

BUG=b:37434548
TEST=unit tests

Change-Id: Id672aa9a143c41047dc9b8cb37e5b059cee32a35
Reviewed-on: https://chromium-review.googlesource.com/905576
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>