trunks: Fix marshalling of size in complex TPM2B_ structs
1. Fix incorrect serializing and parsing of complex TPM2B_ structures in cases
   where the "size" field is zero.
2. Make consistent and enforce the convention that the "size" field for such
   structures is EITHER zero OR sizeof of the enclosed structure (but NOT the
   length of the marshalled representation nor any other value).

In detail, #1 is, as believed, bringing the implementation in conformance to the
specs:

> When the type of the second parameter in the TPM2B_ structure is not BYTE, the
> value of the size field
> shall either be zero indicating that no structure is to be unmarshaled; or it
> shall be identical to the number
> of octets unmarshaled for the second parameter.

(TPM 2.0 Part 2 Chapter 10.4 "Sized Buffers").

So before this CL, the serializing and parsing functions didn't handle the case
of zero size that is decribed in the quote above. Instead, they always behaved
by serializing/parsing the enclosed structure. For example, a full-zeroed
TPM2B_SENSITIVE was serialized as "\0\x6\0\0\0\0\0\0", instead of just "\0\0".
The CL fixes the behavior to produce the latter. (Which allows to pass such
"empty" structures as TPM function arguments when it's required.)

And the change #2 is, basically, about more consistence in the code, because
before this CL one part of code (e.g., the Make_* functions) was putting the
value of sizeof into the size field, while the other part of code (e.g., the
Parse_* functions) was putting the length of the marshalled representation. The
CL makes all of them to use the former.

BUG=chromium:806788
TEST=existing and new unit tests

Change-Id: Ied11b00f2ead53dac345885915b7d073775f9417
Reviewed-on: https://chromium-review.googlesource.com/906930
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>