cryptohome: Cache enrollment ID in attestation database at enrollment
The enrollment ID depends on the device secret and is used for zero
touch enrollment or for re-enrollment. When the device is enrolled
the computed enrollment ID has to be cached in the attestation database.

This allows to avoid problems in a scenario when the device secret is
changed after the device is enrolled. If the enrollment ID would be
computed after that, it would be different from the value used at the
enrollment. Having it stored in attestation database allows to be sure
the enrollment ID remains the same.

In this CL the enrollment ID is computed at the enrollment and stored
in attestation database. When it is requested, we check the value from
the database. If it is present there, then that's the value returned.

BUG=chromium:793326
TEST=cryptohome --action=get_enrollment_id

Change-Id: Ic53d4b1aef73b3983947d082e10406715db65f29
Reviewed-on: https://chromium-review.googlesource.com/906630
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>