trunks: Stricter checks of blob parsing leftovers
This CL throws in a few more checks that the data buffer was completely parsed,
without a single byte staying left, after performing the TPM types parsing.

In other words, this change is effectively adding more assertions on the input
data format.

The expectation is that nothing should break as a result of this change,
assuming that the existing client code calls trunks functions with correct
inputs.

BUG=chromium:806788
TEST=existing and new unit tests

Change-Id: Iaaac11df3619b5ce04ba6465b9d5085a76b75561
Reviewed-on: https://chromium-review.googlesource.com/908553
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>