cryptohome: Signature-protected sealing with TPM2
Implement sealing of secret data to the specified public key, such that
unsealing requires signing a challenge (a random nonce); the implementation is
currently provided only for TPM 2.0.

This sealing is based on the TPM2_PolicySigned functionality of TPM 2.0.

The CL is also aiming at structuring the code so that the same internal APIs
(SignatureSealingBackend and related) can be implemented for TPM 1.2 as well -
but that will happen in future CLs.

Besides mock-based unit tests, also a "live" test is provided, which allows to
test the sealing and unsealing against real TPM's.

BUG=chromium:806788
TEST=unit tests: Tpm2RsaSignatureSecretSealingTest.*;
     running live test on a real device (snappy) via
     cryptohome --action=tpm_live_test

Change-Id: I7c2f2b1230fdb07c625a4becf04c2b78b46f6352
Reviewed-on: https://chromium-review.googlesource.com/914903
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>