authpolicy: Manage machine account password
Manages the machine password in authpolicyd instead of Samba. This is a
necessary first step to change the machine password using kpasswd, which
requires the old password as input (the machine keytab is not enough).
The password change will be done in a future CL.

For newly joined machines the machine password is used to authenticate
the device. However, on existing devices that do not have the machine
password anymore, the keytab file is used just like before.

Previously, Samba stored the machine password secrets.tdb in Samba's
private folder. Since authpolicyd puts that on tmpfs, it was lost on
restart. Alternatively, secrets.tdb could be put on encstateful.
However, this would store not only the machine password on encstateful,
but potentially other sensitive data and it is hard to reason
security-wise. Hence, this approach was rejected.

BUG=chromium:777979
TEST=cros_run_unit_tests --board=amd64-generic --packages authpolicy
      test_that -b ${BOARD} ${DUT_IP} enterprise_AuthPolicyDaemon
      Tested that old clients use keytab instead of password.
      Tested on Chrome OS device with live server.

Change-Id: I090c00d24265109d7fe502eab1d62136d79d47b7
Reviewed-on: https://chromium-review.googlesource.com/914148
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>