vm_tools: concierge: Don't run minijail as pid 1 in namespace
When creating a new pid namespace, minijail will run as pid 1 in that
namespace by default and when it receives a SIGTERM, it exits
immediately.

Unfortunately, this means that every process in that pid namespace in
unceremoniously killed without being given a chance to clean up.  This
is how you get file system corruption in the VM.

Instead, tell minijail not to run as pid 1.  Concierge will take care of
handling SIGTERM and will attempt an orderly shut down of any running
VMs.  Also increase the kill timeout to give VMs time to shut down.

BUG=none
TEST=stop vm_concierge; check /var/log/messages for shutdown RPC
     confirmation from maitre'd

Change-Id: I72b54dc3101eb6625ffca4622c1e1f994ba0ed61
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/923390
Reviewed-by: Mike Frysinger <vapier@chromium.org>