vm_tools: concierge: Fix hang when crosvm does not exit
'crosvm stop' does not return a meaningful exit code and even if it did
the only wait to really be sure that crosvm has exited is to get that
information from the kernel.

Stop checking the return code from 'crosvm stop' and just wait for
crosvm to exit.  In addition, make sure there is always a timeout when
waiting for a child to exit so that concierge doesn't end up in a hung
state.

BUG=none
TEST=manual. have concierge start and stop multiple VMs

Change-Id: Ifcf75a22d52f76040928641e4f0c81c74cd87f8c
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/923391
Reviewed-by: Mike Frysinger <vapier@chromium.org>