chromeos-config: Provide a way to compare FDT and YAML result
We should get the same result regardless of whether we use FDT or YAML
formats for our config, with some exceptions. Add an option to
cros_config_host to validate that this is in fact true. Make the option
on by default, assuming that a config file in the other format can be
found.

NOTE: The exceptions relate to minor differences between the fdt and
yaml config which appear in tests but which will not be used in real
config files. This means that we should catch any differences in the
wild.

BUG=chromium:796983
TEST=FEATURES=test sudo -E emerge  --nodeps chromeos-config-host \
            chromeos-config-tools

Change-Id: Id9ea4684eba0c636f95d464f9d3a61e499fcafe4
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/927008
Reviewed-by: Jason Clinton <jclinton@chromium.org>