cryptohome: Add unit tests for LECredentialManager
Add a few unit tests for the LECredentialManager class.
In order to meaningfully verify class behaviour, we introduce
a FakeLECredentialBackend implementation of the LECredentialBackend
interface, which mimics the expected behaviour on the Cr50 side.

A few things are missing in the fake implementation, which will
be added in subsequent CLs:
- Time stamp support
- Delay schedule support.

Still, this serves as a functioning baseline, from which we can modify
both the unit tests and the FakeLECredentialBackend, as the
corresponding behaviour in the real Cr50 implementation gets finalized.

BUG=chromium:794010, chromium:809718, chromium:809720, chromium:809713
TEST=cros_run_unit_tests passes on eve.

Change-Id: I12ceeff46ed9beba283ce78b43ec971e07430854
Reviewed-on: https://chromium-review.googlesource.com/933081
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>