vm_tools: concierge: increase startup timeout to 30 seconds
If you try extracting a large compressed image (e.g. a 32GiB disk image
compressed with XZ), then 5 seconds can be too low for the VM startup.
Lets increase this to something much higher like 30 seconds.  In the
normal case, we aren't going to hit that as we connect as soon as the
VM is ready.  This timeout only shows up when there is a failure.

BUG=None
TEST=vm still starts up

Change-Id: I2808b1257b1682c1f665bf95ebfc08ecf6cfbf50
Reviewed-on: https://chromium-review.googlesource.com/928846
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>