power: New Dark Resume Logic.
Use wakeup_count to identify if the wake is a dark resume.
If the wake is triggered by the user, then the system does a full
resume. Otherwise, the system does a dark resume. We still keep the
existing logic(renamed to legacy_dark_resume.cc) as is enabled on the
link.

BUG=b:68944516
TEST=None

Change-Id: I165455ddd6e2c914b15c207550d9f8111e4e0708
Reviewed-on: https://chromium-review.googlesource.com/932760
Commit-Ready: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>