chromeos-config: Split out identity impl into its own class
We have a separate header file for identity, but at present the
implementation is mixed up in the main cros_config.cc file. This is a bit
confusing. It seems cleaner to split it into a separate class.

Create a new CrosConfigIdentity class and move methods related to identity
into it.

BUG=chromium:796983
TEST=FEATURES=test sudo -E emerge  --nodeps chromeos-config-host \
	chromeos-config-tools

Change-Id: I0c5fc2ca2ccaf485345d69297175aeeda517c688
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/936474
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>