chromeos-config: Rename 'offset' to node
The word 'offset' is a flat-device-tree concept. In yaml we cannot access
nodes by their offset in the file. The offset would also not be useful if
we used a live device tree (e.g. like Linux).

Use the term 'node' instead which works for both yaml and device tree.

BUG=chromium:796983
TEST=FEATURES=test sudo -E emerge --nodeps chromeos-config-host \
	chromeos-config-tools

Change-Id: Idcd4865a78bff8e487b1278f176d1d3bb83516f9
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/937389
Reviewed-by: C Shapiro <shapiroc@google.com>