chromeos-config: Format code with clang-format
The 'cros lint' pre-submit check currenty passes, but clang-format makes a
number of suggested changes to the code. It is easier to follow these
rules since it reduces the burden of getting the code formating correct.

Update the code to match clang-format's view of the situation.

It is possible to enable clang-format in the pre-submit, but this does
work when using 'repo upload' outside the chroot (I get a 'permission
denied' error).

BUG=chromium:798771
TEST=clang-format -i -style=file $(find . -name '*.h' -o -name '*.cc')

Change-Id: I2e731e8d1887c4d072bd6844ece991d79de829fa
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/937390
Reviewed-by: C Shapiro <shapiroc@google.com>