chromeos-config: Move FDT into a separate class
We want to use both the FDT and yaml configs and check that they return
the same result for each query. In preparation for this, move the main
implementation into a new CrosConfigImpl class, and add a CrosConfigFdt
subclass which implements the FDT-specific pieces.

Most of the code from cros_config.cc moves to cros_config_impl.cc since
we will now have implementations specific to the config file. The code
in lookup.cc moves to cros_config_fdt.cc since is device-tree-specific.
You should be able to use meld to compare files in the previous commit
with their corresponding files.

The CrosConfig class is mostly just a veneer, calling into CrosConfigImpl
subclasses to do the real work.

After this we can add a YAML subclass to implement that, and compare the
results.

NOTE: This is all in one large commit to avoid grumbling.

BUG=chromium:796983
TEST=FEATURES=test sudo -E emerge --nodeps chromeos-config-host \
	chromeos-config-tools

Change-Id: Ibd2aa61e0fbad425a437adad622bdbe4f1f9502a
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/937392
Reviewed-by: Jason Clinton <jclinton@chromium.org>