cryptohome: Add encstateful NVRAM space for TPM1
This adds support for the TPM1 system key loading code to get the system key
from a dedicated NVRAM space (as opposed to using the system salt in the
lockbox space). This has a number of benefits, including the ability to
configure a TPM-backed system key when the lockbox space doesn't exist or is
still locked after TPM clear.

BUG=None
TEST=compiles and passes tests

Change-Id: Ib1bb292221b4a2776e31a0a767b26eb9fad46cb5
Reviewed-on: https://chromium-review.googlesource.com/937708
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>