rockchip: Fix PostProcThread crash
When supporting recoverable request error, we assumed metadata it the
last returned result, so we force returning all buffers before that to
avoid buffer leak and keep the results and request error message in sync.

But that is no longer true after we made the post processing
asynchronous in:
4c4b04f5 rockchip: process caputre stream in independent thread when video snapshot

So the PostProcThread may crash when trying to access resources which are
already freed before metadata done.

Adjust request error handling to avoid that.

BUG=b:72365664, b:73868843
TEST=test on my scarlet:
1/ pass CTS RobustnessTest#testMandatoryOutputCombinations
2/ fake random request error in OutputFramework's prepareRun and processFrame

Change-Id: I044f58f89b6800b79cb6b2252f6caab8c25f7d1e
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/939122
Reviewed-by: Hung-yu Wu <hywu@chromium.org>