cryptohome: Key preservation support in mount-encrypted
Allow mount-encrypted to preserve the encryption key used before the
TPM was cleared and hence retain the encrypted file system across TPM
reset. This is triggered by a flag file. If the file is present when
the TPM is clear and missing a valid system key, we try to load the
previous system key, use it to decrypt the previous encryption key,
then generate a fresh system key to rewrap the encryption key. The
intended use of this is TPM firmware updates, which require a TPM
clear to install. We'd like to allow carrying over the stateful file
system across TPM firmware update installation.

BUG=chromium:788719
TEST=New unit tests.

Change-Id: I675f87818bb1a41ce806a602b734681e613e116c
Reviewed-on: https://chromium-review.googlesource.com/944062
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>