cryptohome: Skip AIK loading for TPM 1.2 delegate creation
Remove loading of the Attestation Identity Key before creating the TPM 1.2
delegate - as, presumably, this was a no-op that didn't affect the delegate
creation in any way. The AIK used to be simply loaded and not passed anywhere
explicitly; the Trousers' delegate creation procedures also don't seem to depend
on whether the AIK is loaded or not.

Besides removing unnecessary code, this change should also simplify writing TPM
live tests that need to work with delegates.

BUG=chromium:806788
TEST=existing tests;
     the "platform_Attestation" autotest on a TPM 1.2 device (e.g. leon)

Change-Id: I49d679f5ac093af6f47e162eee2bff2347315c5c
Reviewed-on: https://chromium-review.googlesource.com/939628
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>