power: powerd_suspend: always write hwclock result, when requested
When a test is asking us to stash hwclock output to
/run/power_manager/root/hwclock-on-resume, we only do that if hwclock
exits successfully. But this can leave stale results, which is worse
than just having error logs. (And in fact, hwclock just gives empty
output on errors anyway, unless run with --debug.)

Help to ensure that tests don't process stale data writing output
in both success and error cases.

There are accompanying changes in http://crosreview.com/940601 to
improve the test itself, but we might as well do it here too.

BUG=chromium:815249
TEST=power_SuspendStress with misbehaving RTC

Change-Id: I52b1c69aa114ce583c09c3ed5b9d84bc618d301d
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/940761
Reviewed-by: Dan Erat <derat@chromium.org>