vm_tools: Initial version of garcon
Design doc for this is here: goto/garconbridge

This adds a simple version of garcon to start with that just notifies
the host that garcon executed and container startup is complete.

This still needs to have the iptable rules changed to allow connection
from the container and also bind-mount garcon into the container.

BUG=none
TEST=Verified host receives the startup message from garcon
CQ-DEPEND=CL:954019

Change-Id: I16028b5e3a3d42e23bd34d8df8e4c0fb13686fa8
Reviewed-on: https://chromium-review.googlesource.com/954018
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>