power: ignore power button release only when matching a press
powerd should ignore a power button 'up' event only if it has ignored
the matching 'down' event, else the UI might see an artificial long power
button event and shutdown the machine.

If there is a delay between when a website displays the physical
presence request and when the u2fd daemon receives this same request,
the user might have pressed the power button before we are even aware of
the fact that we should ignore the next button cycle and trigger the bad
case described above.
As a simple workaround, let's ignore the button release too in this case.
I have personally observed that this delay has dramatically increased
recently (for an unknown reason), the users are more likely to see this
issue.

BUG=b:73349998
TEST=On Eve, manually do login cycles as fast as I can.
TEST=the updated InputEventHandlerTest.IgnorePowerButtonPresses unittest
passes with the rest of the patch and fails without.

Change-Id: If156429f430c360bcaac81d00115afec324f7250
Reviewed-on: https://chromium-review.googlesource.com/948845
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>