hammerd: init script forwards minijail exit status
In the current version of the hammerd init script, the exit status
from minijail0 is acted on correctly (write or delete tag file),
but that exit status is not forwarded on to upstart correctly.
This prevents service failure crash reports from showing the
correct exit status.

This CL modifies the init script to correctly forward the exit
status to upstart.  It also checks for failure to write into the
WRITE_SYSFS_PATH tag file.

BUG=b:69892693
TEST=Run `start hammerd` on DUT, and check /var/log/messages for one of
     - INFO minijail0[11215]: libminijail[11215]: child process 11216 exited with status 11
       WARNING kernel: [ 3391.302614] init: hammerd main process (11214) terminated with status 11
     - NOTICE hammerd[11502]: FAILED to write into WRITE_SYSFS_PATH tag file
       WARNING kernel: [ 3558.213973] init: hammerd main process (11497) terminated with status 1

Change-Id: Ifd4c2a53a3bc88a972949c54a649d94b0e01198c
Reviewed-on: https://chromium-review.googlesource.com/952119
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>