chromeos-config: Add command for firmware combinations
chromeos-bootimage needs to look up the relevant coreboot/depthcharge
pairs when assembling a bootimage. coreboot needs to do something
similar for building coreboot/EC pairs.  The commands that exist in
cros_config_host today aren't sufficient for this, so this commit
introduces a new command that returns all of the sets of build targets
for a given config.

TEST=run_tests.sh
TEST=python -m cros_config_host.cros_config_host get-firmware-build-combinations coreboot,ec,depthcharge for CL:*583896
BUG=chromium:797853

Change-Id: I8410a479e15a2aeaeb42d0b487d31a4fac057717
Reviewed-on: https://chromium-review.googlesource.com/953124
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>