cryptohome: Generate reset seed for Credentials
This patch generates a reset seed in a VaultKeyset.
This reset seed is then used to generate a reset secret
which can be used to unlock a Low Entropy (LE) credential
that has been locked out.

The reset seed is generated as follows:
During Homedirs::AddKeyset() we check if the existing VaultKeyset
has a wrapped_reset_seed().

If it doesn't, we generate a random 32-byte reset seed in the
VaultKeyset, and then re-encrypt the same credential, so that
it is persisted to disk, and made part of the SerializedVaultKeyset.

If a reset seed is already present, we don't do anything.

The rest of the AddKeyset() function proceeds as before.

The reset_seed will then be used by EncryptLECredential() to generate
a reset_secret with which the LE credential is stored.

BUG=chromium:794010,chromium:809723
TEST=cros_run_unit_tests passes.

Change-Id: I2bd025ef672294381cc7baf7503caaddb2bd68d7
Reviewed-on: https://chromium-review.googlesource.com/955283
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>