authpolicy: Add a way to disable log anonymizer
Adds a flag to disable the log anonymizer. This is strictly a debug only
flag. It can only be set on a Chromebook in dev mode and writeable
rootfs with

  echo '{"disable_anonymizer":true}' > /etc/authpolicyd_flags

The crosh command authpolicy_debug never sets it since this command is
available for everyone and in non-dev mode.

BUG=chromium:776675
TEST=cros_run_unit_tests --board=amd64-generic --packages authpolicy
     Tested on device

Change-Id: I99106aa7d1efa6b8c8cfb6bd8c3411a2f7fd9606
Reviewed-on: https://chromium-review.googlesource.com/955643
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>