libbrillo: Support adding/removing interfaces and their properties.
D-Bus specification allows interfaces to be added or removed even after
an object is exported. This CL adds the support to libbrillo's D-Bus
library to do this.

This is needed by the new Bluetooth daemon which is going to impersonate
BlueZ's ObjectManager and BlueZ may add/remove interfaces of an object
dynamically.

BUG=chromium:812468
TEST=Tested with Bluetooth daemon

Change-Id: Ia72735bfd28082cf6c43d1e774beb912fa60ec2d
Reviewed-on: https://chromium-review.googlesource.com/956745
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>