bluetooth: Move DaemonTest to SuspendManagerTest
The current DaemonTest exercises detailed code paths of SuspendManager
and this test is more suitable to be a unit test of SuspendManager. This
prepares for the upcoming CL of adding BlueZ forwarding in Daemon and
DaemonTest should then contain the integration tests exercising both
SuspendManager and BlueZ forwarding.

BUG=chromium:812468
TEST=Ran the moved unit test

Change-Id: Ibc91ebe51ade7767907d2118717fe409fb246094
Reviewed-on: https://chromium-review.googlesource.com/958129
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>