libbrillo: Expose DBusObject of ExportedObjectManager.
ExportedObjectManager subscribes to signals on interface
org.freedesktop.DBus.ObjectManager. In D-Bus forwarding case, it may
receive expected signals from the forwarded service about
InterfacesAdded and InterfacesRemoved but it would be logged as
unexpected method calls and pollute the logs. There needs to be a way to
freely attach method handlers to the expected methods so we can handle
them instead of unnecessarily polluting the logs.

BUG=chromium:812468
TEST=Tested with Bluetooth daemon D-Bus forwarding

Change-Id: Ifb26a2d7ce800fde4311b18efe36ab5110f0dae4
Reviewed-on: https://chromium-review.googlesource.com/958325
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>