chromeos-config: Just log yaml failures
When yaml queries don't return the correct result, we write out to the
log and return an error.

However there is still at least one infidelity to resolve. While we do
this, just log if the yaml results are incorrect, or the yaml file is
missing entirely.

BUG=chromium:813855, b:74508666
TEST=manually test on reef that this failure just logs:
cros_config /audio/main ucm-suffix

Change-Id: I07a8b76774a43f6b04f42104ca864546bec351cd
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/959050
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>