cryptohome: Service method for LockToSingleUserMountUntilReboot added
To implement security feature that would force the device to reboot
after ARC session is over, Chrome will need to disable login until reboot
of the device is done. This will be achieved by extending a PCR that is
bound to cryptohome key. The method is added in the service to be
available to be called from Chrome. The command line added in cryptohome
is optional and used only for testing.

BUG=chromium:812165
TEST=cryptohome --action=lock_to_single_user_mount_until_reboot --user=<email>
CQ-DEPEND=CL:1458577

Change-Id: I121d19106ee04fde833111ba9de87842bff7dfde
Reviewed-on: https://chromium-review.googlesource.com/958866
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>